Moved game logo code into own class
